+++ R/CreateArgFunctions.R
@@ -14,7 +14,8 @@
 #'                                         negatives).
 #' @param prevalenceCohortId               The number of the cohort definition id to determine
 #'                                         the disease prevalence.
-#' @param xSpecCohortSize                  The recommended xSpec sample size to use in model (default = 5000)
+#' @param caseCohortId                     The number of the cohort definition id to determine cases in the evaluation cohort
+#' @param caseFirstOccurrenceOnly          Set to true if only the first occurrence per subject in the case cohort is to be used
 #' @param covariateSettings                A covariateSettings object as generated
 #'                                         using createCovariateSettings().
 #' @param modelPopulationCohortId          The number of the cohort to be used as a base population for
@@ -65,6 +66,8 @@ createCreateEvaluationCohortArgs <- function(xSpecCohortId,
                                              daysFromxSpec = 0,
                                              xSensCohortId = prevalenceCohortId,
                                              prevalenceCohortId,
+                                             caseCohortId = prevalenceCohortId,
+                                             caseFirstOccurrenceOnly = TRUE,
                                              xSpecCohortSize = 5000,
                                              covariateSettings = createDefaultCovariateSettings(
                                                excludedCovariateConceptIds = c(),
@@ -79,8 +82,8 @@ createCreateEvaluationCohortArgs <- function(xSpecCohortId,
                                              exclusionEvaluationCohortId = 0,
                                              exclusionEvaluationDaysFromStart = 0,
                                              exclusionEvaluationDaysFromEnd = 0,
-                                             minimumOffsetFromStart = 365,
-                                             minimumOffsetFromEnd = 365,
+                                             minimumOffsetFromStart = 0,
+                                             minimumOffsetFromEnd = 30,
                                              modelBaseSampleSize = 25000,
                                              baseSampleSize = 2e+06,
                                              lowerAgeLimit = 0,