+++ R/CreateEvalCohort.R
@@ -22,6 +22,8 @@
                                     xSpecCohortId,
                                     xSensCohortId,
                                     prevalenceCohortId,
+                                    caseCohortId,
+                                    caseFirstOccurrenceOnly,
                                     cdmDatabaseSchema,
                                     cohortDatabaseSchema,
                                     cohortTable,
@@ -55,12 +57,17 @@
                                     excludeModelFromEvaluation = FALSE,
                                     randomVisitTable = "",
                                     savePlpData = FALSE) {
-  if (savePlpData == TRUE) {
-    evaluationCohortPlpDataFileName <- file.path(outFolder, sprintf("evaluationCohortPlpData_%s", evaluationCohortId))
-    if (file.exists(evaluationCohortPlpDataFileName)) {
-      stop("The savePlpData argument is set to TRUE, but ", evaluationCohortPlpDataFileName, " already exists")
-    }
-  }
+
+  tempTableCreated <- FALSE
+  evaluationCohortPlpDataFileName <- file.path(outFolder, sprintf("evaluationCohortPlpData_%s", evaluationCohortId))
+
+  # if (savePlpData == TRUE) {
+  #   evaluationCohortPlpDataFileName <- file.path(outFolder, sprintf("evaluationCohortPlpData_%s", evaluationCohortId))
+  #   if (file.exists(evaluationCohortPlpDataFileName)) {
+  #     stop("The savePlpData argument is set to TRUE, but ", evaluationCohortPlpDataFileName, " already exists")
+  #   }
+  # }
+
   evaluationCohortFileName <- file.path(outFolder, sprintf("evaluationCohort_%s.rds", evaluationCohortId))
   if (file.exists(evaluationCohortFileName)) {
     ParallelLogger::logInfo("Skipping creation of ", evaluationCohortFileName, " because it already exists")
@@ -99,6 +106,7 @@
 
       tryCatch(
         {
+          tempTableCreated <- TRUE
           insertTable(
             connection = connection,
             databaseSchema = workDatabaseSchema,
@@ -134,57 +142,61 @@
       )
     } else { # otherwise create the evaluation cohort from an sql query
 
-      sqlFilename <- "CreateCohortsAcuteEvaluation.sql"
+      if (!file.exists(evaluationCohortPlpDataFileName)) { #only create evaluation cohort if data file does not exist
+        ParallelLogger::logInfo("Using Cohort Id ", caseCohortId, " to designate cases.")
+        sqlFilename <- "CreateCohortsAcuteEvaluation.sql"
 
-      if (inclusionEvaluationCohortId != 0) {
-        ParallelLogger::logInfo("Creating evaluation cohort subjects using visits from cohort Id: ", inclusionEvaluationCohortId)
+        if (inclusionEvaluationCohortId != 0) {
+          ParallelLogger::logInfo("Creating evaluation cohort subjects using visits from cohort Id: ", inclusionEvaluationCohortId)
+        }
+        if (exclusionEvaluationCohortId != 0) {
+          ParallelLogger::logInfo("Creating evaluation cohort subjects excluding visits from cohort Id: ", exclusionEvaluationCohortId)
+        }
+
+        if (randomVisitTable != "") {
+          ParallelLogger::logInfo("Creating evaluation cohort subjects using supplied random visit table: ",
+                                  paste0(workDatabaseSchema, ".", randomVisitTable))
+        }
+
+        sql <- SqlRender::loadRenderTranslateSql(
+          sqlFilename = sqlFilename,
+          packageName = "PheValuator",
+          dbms = connectionDetails$dbms,
+          cdm_database_schema = cdmDatabaseSchema,
+          cohort_database_schema = cohortDatabaseSchema,
+          cohort_database_table = cohortTable,
+          tempEmulationSchema = tempEmulationSchema,
+          x_spec_cohort = xSpecCohortId,
+          caseCohortId = caseCohortId,
+          caseFirstOccurrenceOnly= caseFirstOccurrenceOnly,
+          work_database_schema = workDatabaseSchema,
+          test_cohort = testCohort,
+          exclCohort = 0,
+          ageLimit = lowerAgeLimit,
+          upperAgeLimit = upperAgeLimit,
+          gender = gender,
+          race = race,
+          ethnicity = ethnicity,
+          startDate = startDate,
+          endDate = endDate,
+          baseSampleSize = format(baseSampleSize, scientific = FALSE),
+          xSpecSampleSize = 5,
+          inclusionEvaluationCohortId = inclusionEvaluationCohortId,
+          inclusionEvaluationDaysFromStart = inclusionEvaluationDaysFromStart,
+          inclusionEvaluationDaysFromEnd = inclusionEvaluationDaysFromEnd,
+          exclusionEvaluationCohortId = exclusionEvaluationCohortId,
+          exclusionEvaluationDaysFromStart = exclusionEvaluationDaysFromStart,
+          exclusionEvaluationDaysFromEnd = exclusionEvaluationDaysFromEnd,
+          minimumOffsetFromStart = minimumOffsetFromStart,
+          minimumOffsetFromEnd = minimumOffsetFromEnd,
+          randomVisitTable = randomVisitTable,
+          visitLength = visitLength,
+          visitType = c(visitType)
+        )
+        ParallelLogger::logInfo("Creating evaluation cohort on server from sql")
+        DatabaseConnector::executeSql(connection = connection, sql)
+        tempTableCreated <- TRUE
       }
-      if (exclusionEvaluationCohortId != 0) {
-        ParallelLogger::logInfo("Creating evaluation cohort subjects excluding visits from cohort Id: ", exclusionEvaluationCohortId)
-      }
-
-      if (randomVisitTable != "") {
-        ParallelLogger::logInfo("Creating evaluation cohort subjects using supplied random visit table: ",
-                                paste0(workDatabaseSchema, ".", randomVisitTable))
-      }
-
-      sql <- SqlRender::loadRenderTranslateSql(
-        sqlFilename = sqlFilename,
-        packageName = "PheValuator",
-        dbms = connectionDetails$dbms,
-        cdm_database_schema = cdmDatabaseSchema,
-        cohort_database_schema = cohortDatabaseSchema,
-        cohort_database_table = cohortTable,
-        tempEmulationSchema = tempEmulationSchema,
-        x_spec_cohort = xSpecCohortId,
-        prevalenceCohortId = prevalenceCohortId,
-        work_database_schema = workDatabaseSchema,
-        test_cohort = testCohort,
-        exclCohort = 0,
-        ageLimit = lowerAgeLimit,
-        upperAgeLimit = upperAgeLimit,
-        gender = gender,
-        race = race,
-        ethnicity = ethnicity,
-        startDate = startDate,
-        endDate = endDate,
-        baseSampleSize = format(baseSampleSize, scientific = FALSE),
-        xSpecSampleSize = 100,
-        inclusionEvaluationCohortId = inclusionEvaluationCohortId,
-        inclusionEvaluationDaysFromStart = inclusionEvaluationDaysFromStart,
-        inclusionEvaluationDaysFromEnd = inclusionEvaluationDaysFromEnd,
-        exclusionEvaluationCohortId = exclusionEvaluationCohortId,
-        exclusionEvaluationDaysFromStart = exclusionEvaluationDaysFromStart,
-        exclusionEvaluationDaysFromEnd = exclusionEvaluationDaysFromEnd,
-        minimumOffsetFromStart = minimumOffsetFromStart,
-        minimumOffsetFromEnd = minimumOffsetFromEnd,
-        randomVisitTable = randomVisitTable,
-        visitLength = visitLength,
-        visitType = c(visitType)
-      )
-
-      ParallelLogger::logInfo("Creating evaluation cohort on server from sql")
-      DatabaseConnector::executeSql(connection = connection, sql)
     }
 
     # will only use the covariates with non-zero betas
@@ -197,7 +209,6 @@
       covariateSettings[[listUp]]$includedCovariateIds <- c(lrNonZeroCovs)
     }
 
-    evaluationCohortPlpDataFileName <- file.path(outFolder, sprintf("evaluationCohortPlpData_%s", evaluationCohortId))
     if (file.exists(evaluationCohortPlpDataFileName)) {
       ParallelLogger::logInfo("Getting evaluation cohort data from existing folder")
       plpData <- PatientLevelPrediction::loadPlpData(evaluationCohortPlpDataFileName)
@@ -276,77 +287,141 @@
     # apply the model to the evaluation cohort
     appResults <- NULL
 
-
-    # appResults$prediction <- PatientLevelPrediction::predictPlp(
-    #   plpModel = lrResults$model, plpData = plpData,
-    #   population = population
-    # )
-
     ####################
-    ParallelLogger::logInfo("Applying predictive model to evaluation cohort using rapid method")
+    ParallelLogger::logInfo("Applying predictive model to evaluation cohort using predictPlp method")
 
     startTime <- Sys.time()
 
-    # extract the normalization and stuff
-    covs <- lrResults$model$covariateImportance[lrResults$model$covariateImportance$covariateValue !=0,]
-
-    intercept <- lrResults$model$model$coefficients$betas[1]
-
-    covariate <- data.frame(
-      coefficient = covs$covariateValue,
-      covariateId = covs$covariateId)
-
-    norm <- as.data.frame(lrResults$model$preprocessing$tidyCovariates$normFactors)
-    covNorm <- merge(covariate, norm, by = 'covariateId', all.x = T)
-
-    allCovData <- as.data.frame(plpData$covariateData$covariates)
-
-    #allCovData <- merge(allCovData, covNorm, by = 'covariateId', all.x = T)
-
-    allCovData <- dplyr::left_join(allCovData, covNorm, by = 'covariateId')
-
-    allCovData$coeffValue <- allCovData$covariateValue*allCovData$coefficient/allCovData$maxValue
-    allCovData$coeffValue[is.na(allCovData$coeffValue)] <- 0
-
-    sumData <- aggregate(allCovData$coeffValue, by=list(Category=allCovData$rowId), FUN=sum)
-
-    sumData$x[is.na(sumData$x)] <- 0
-    sumData$x <- sumData$x + intercept
-    sumData$probability <- 1/(1+exp(-1*sumData$x))
-    sumData$probability <- round(sumData$probability, digits = 3)
-    names(sumData)[1] <- "rowId"
-    names(sumData)[3] <- "value"
-
-    sumData <- sumData[,c(1,3)]
-    appResults$prediction <- merge(population, sumData, by = 'rowId')
+    appResults$prediction <- PatientLevelPrediction::predictPlp(
+      plpModel = lrResults$model, plpData = plpData,
+      population = population
+    )
 
     ParallelLogger::logInfo("Time to apply model was ", round(difftime(Sys.time(), startTime, units = c("mins")), digits = 3), " min.")
+    allCovData <- as.data.frame(plpData$covariateData$covariates)
+
+    ####################
+    #
+    #     ParallelLogger::logInfo("Applying predictive model to evaluation cohort using rapid method")
+    #
+    #     startTime <- Sys.time()
+    #
+    #     #extract the normalization and stuff
+    #     covs <- lrResults$model$covariateImportance[lrResults$model$covariateImportance$covariateValue !=0,]
+    #
+    #     intercept <- lrResults$model$model$coefficients$betas[1]
+    #
+    #     covariate <- data.frame(
+    #       coefficient = covs$covariateValue,
+    #       covariateId = covs$covariateId)
+    #
+    #     norm <- as.data.frame(lrResults$model$preprocessing$tidyCovariates$normFactors)
+    #     covNorm <- merge(covariate, norm, by = 'covariateId', all.x = T)
+    #
+    #     allCovData <- as.data.frame(plpData$covariateData$covariates)
+    #
+    #     #allCovData <- merge(allCovData, covNorm, by = 'covariateId', all.x = T)
+    #
+    #     allCovData <- dplyr::left_join(allCovData, covNorm, by = 'covariateId')
+    #
+    #     allCovData$coeffValue <- allCovData$covariateValue*allCovData$coefficient/allCovData$maxValue
+    #     allCovData$coeffValue[is.na(allCovData$coeffValue)] <- 0
+    #
+    #     sumData <- aggregate(allCovData$coeffValue, by=list(Category=allCovData$rowId), FUN=sum)
+    #
+    #     sumData$x[is.na(sumData$x)] <- 0
+    #     sumData$x <- sumData$x + intercept
+    #     sumData$probability <- 1/(1+exp(-1*sumData$x))
+    #     sumData$probability <- round(sumData$probability, digits = 3)
+    #     names(sumData)[1] <- "rowId"
+    #     names(sumData)[3] <- "value"
+    #
+    #     sumData <- sumData[,c(1,3)]
+    #     appResults$prediction <- merge(population, sumData, by = 'rowId')
+    #
+    #     ParallelLogger::logInfo("Time to apply model was ", round(difftime(Sys.time(), startTime, units = c("mins")), digits = 3), " min.")
     ###################
 
-    appResults$prediction$value <- round(appResults$prediction$value, digits = 3)
+    appResults$prediction$value <- round(appResults$prediction$value, digits = 2)
 
     pred <- appResults$prediction
 
-    # pull in the prevalence cohort to mark the cases
+    # pull in the designated case cohort to mark the cases
     sql <- SqlRender::loadRenderTranslateSql("GetComparisonCohort.sql",
                                              packageName = "PheValuator",
                                              dbms = connectionDetails$dbms,
                                              cohort_database_schema = cohortDatabaseSchema,
                                              cohort_table = cohortTable,
-                                             cohort_id = prevalenceCohortId
-    )
+                                             cohort_id = caseCohortId,
+                                             caseFirstOccurrenceOnly = caseFirstOccurrenceOnly,
+                                             inclusionEvaluationCohortId = inclusionEvaluationCohortId,
+                                             inclusionEvaluationDaysFromStart = inclusionEvaluationDaysFromStart,
+                                             inclusionEvaluationDaysFromEnd = inclusionEvaluationDaysFromEnd)
+
     sql <- SqlRender::translate(sql, connectionDetails$dbms)
     comparisonPopn <- DatabaseConnector::querySql(connection = connection, sql = sql, snakeCaseToCamelCase = TRUE)
-    # add the start dates from the comparison cohort to the evaluation cohort to be able to create
+    # add the start/end dates from the comparison cohort to the evaluation cohort to be able to create
     # dataframe of TP, FP, TN, FN
     finalPopn <- merge(pred, comparisonPopn, all.x = TRUE)
+
+    if(inclusionEvaluationCohortId != 0) {
+      #get inclusion cohort
+      sql <- paste0("select co.subject_id, co.cohort_start_date inclusion_cohort_start_date, co.cohort_end_date inclusion_cohort_end_date from ",
+                    cohortDatabaseSchema, ".", cohortTable, " co ",
+                    " where cohort_definition_id = ", inclusionEvaluationCohortId)
+
+      inclusionCohort <- renderTranslateQuerySql(connection = connect(connectionDetails), sql, snakeCaseToCamelCase = TRUE)
+      inclusionCohort <-  inclusionCohort[inclusionCohort$subjectId %in% c(finalPopn$subjectId),]
+      #merge with finalPopn
+      finalPopn <- merge(finalPopn, inclusionCohort, all.x = TRUE)
+      #remove those where cohort_start_date is not between inclusion start and end
+      finalPopn <- finalPopn[finalPopn$cohortStartDate >= (finalPopn$inclusionCohortStartDate + inclusionEvaluationDaysFromStart) &
+                               finalPopn$cohortStartDate <= (finalPopn$inclusionCohortEndDate + inclusionEvaluationDaysFromEnd), ]
+    } else {
+      #set inclusion start to be cohort_start_date - days from obs start
+      finalPopn$inclusionCohortStartDate <- as.Date(finalPopn$cohortStartDate) - finalPopn$daysFromObsStart
+      #set inclusion end to be cohort_start_date + days to obs end
+      finalPopn$inclusionCohortEndDate <- as.Date(finalPopn$cohortStartDate) + finalPopn$daysToObsEnd
+    }
+
+    #calculate raw time at risk
+    #first for the cases based on inclusion in the comparison cohort
+    #time from the start of subject's in inclusion cohort to time to event
+    finalPopn$rawTar <- 0
+    finalPopn$rawTar[!is.na(finalPopn$comparisonCohortStartDate)] <-
+      as.numeric(difftime(finalPopn$cohortStartDate[!is.na(finalPopn$comparisonCohortStartDate)],
+                          finalPopn$inclusionCohortStartDate[!is.na(finalPopn$comparisonCohortStartDate)], units = "days"))
+
+    #then the non-cases based on non-inclusion in the comparison cohort
+    #time from the start of subject's in inclusion cohort to time to end of subject's in inclusion cohort
+    finalPopn$rawTar[is.na(finalPopn$comparisonCohortStartDate)] <-
+      as.numeric(difftime((finalPopn$inclusionCohortEndDate[is.na(finalPopn$comparisonCohortStartDate)] + inclusionEvaluationDaysFromEnd),
+                          (finalPopn$inclusionCohortStartDate[is.na(finalPopn$comparisonCohortStartDate)] + inclusionEvaluationDaysFromStart), units = "days"))
+
+    #estimated tar = (P(case) * time to event) + (P(not a case) * tar)
+    finalPopn$estimatedTar <- as.numeric((finalPopn$value * difftime(finalPopn$cohortStartDate,
+                                                                     (finalPopn$inclusionCohortStartDate + inclusionEvaluationDaysFromStart), units = "days")) +
+      ((1 - finalPopn$value) * difftime((finalPopn$inclusionCohortEndDate + inclusionEvaluationDaysFromEnd),
+                                        (finalPopn$inclusionCohortStartDate + inclusionEvaluationDaysFromStart), units = "days")))
 
     #determine TP, FP, TN, FN
     fullTestCases <- NULL
     testCases <- finalPopn[is.na(finalPopn$comparisonCohortStartDate) & finalPopn$outcomeCount == 0 & finalPopn$value > 0.8,]
     testCases <- testCases[order(-testCases[,14]),][1:falsePositiveNegativeSubjects,]
-    testCases$type <- "FN"
+    testCases$type <- "FN_Hi"
     fullTestCases <- testCases[!is.na(testCases$subjectId),]
+
+    testCases <- finalPopn[is.na(finalPopn$comparisonCohortStartDate) & finalPopn$outcomeCount == 0 & finalPopn$value > 0.5
+                           & finalPopn$value < 0.6,]
+    testCases <- testCases[order(-testCases[,14]),][1:falsePositiveNegativeSubjects,]
+    testCases$type <- "FN_Med"
+    fullTestCases <- rbind(fullTestCases, testCases[!is.na(testCases$subjectId),])
+
+    testCases <- finalPopn[is.na(finalPopn$comparisonCohortStartDate) & finalPopn$outcomeCount == 0 & finalPopn$value > 0.2
+                           & finalPopn$value < 0.3,]
+    testCases <- testCases[order(-testCases[,14]),][1:falsePositiveNegativeSubjects,]
+    testCases$type <- "FN_Lo"
+    fullTestCases <- rbind(fullTestCases, testCases[!is.na(testCases$subjectId),])
 
     testCases <- finalPopn[!is.na(finalPopn$comparisonCohortStartDate) & finalPopn$outcomeCount == 0 & finalPopn$value < 0.1,]
     testCases <- testCases[order(testCases[,14]),][1:falsePositiveNegativeSubjects,]
@@ -373,41 +448,42 @@
         covs <- merge(covs, modelCovs, by="covariateId", all.x = TRUE)
         covs <- covs[!(is.na(covs$analysisId)),]
 
-        covs$subjectId <- fullTestCases$subjectId[[subjectUp]]
-        covs$type <- fullTestCases$type[[subjectUp]]
+        if(nrow(covs) > 0) {
+          covs$subjectId <- fullTestCases$subjectId[[subjectUp]]
+          covs$type <- fullTestCases$type[[subjectUp]]
 
-        if(nrow(subjectCovariates) == 0) {
-          subjectCovariates <- covs[!is.na(covs$subjectId),]
-        } else {
-          subjectCovariates <- rbind(subjectCovariates, covs[!is.na(covs$subjectId),])
+          if(nrow(subjectCovariates) == 0) {
+            subjectCovariates <- covs[!is.na(covs$subjectId),]
+          } else {
+            subjectCovariates <- rbind(subjectCovariates, covs[!is.na(covs$subjectId),])
+          }
         }
       }
     } else {
       ParallelLogger::logInfo("No test cases found")
     }
 
-
     # save the full data set to the model
     appResults$prediction <- finalPopn
 
     #create diagnostics to later assess analysis performance
     count30And70pct <- round(sum(appResults$prediction$value[appResults$prediction$value >= 0.3 & appResults$prediction$value <= 0.7 &
-                                                               appResults$prediction$outcomeCount == 0]), 0)
-    prop30And70pct <- round(count30And70pct/sum(appResults$prediction$value), 3)
+                                                               appResults$prediction$outcomeCount == 0], na.rm = TRUE), 0)
+    prop30And70pct <- round(count30And70pct/sum(appResults$prediction$value, na.rm = TRUE), 3)
 
     count0And1pct <- round(sum(appResults$prediction$value[appResults$prediction$value >= 0 & appResults$prediction$value <= 0.01 &
-                                                             appResults$prediction$outcomeCount == 0]), 0)
-    prop0And1pct <- round(count0And1pct/sum(appResults$prediction$value), 3)
+                                                             appResults$prediction$outcomeCount == 0], na.rm = TRUE), 0)
+    prop0And1pct <- round(count0And1pct/sum(appResults$prediction$value, na.rm = TRUE), 3)
 
     countGT80pct <- round(sum(appResults$prediction$value[appResults$prediction$value >= 0.8 &
-                                                            appResults$prediction$outcomeCount == 0]), 0)
-    propGT80pct <- round(countGT80pct/sum(appResults$prediction$value), 3)
+                                                            appResults$prediction$outcomeCount == 0], na.rm = TRUE), 0)
+    propGT80pct <- round(countGT80pct/sum(appResults$prediction$value, na.rm = TRUE), 3)
 
     predictionCases <- appResults$prediction[appResults$prediction$outcomeCount == 0 & !(is.na(appResults$prediction$comparisonCohortStartDate)),]
     predictionNonCases <- appResults$prediction[appResults$prediction$outcomeCount == 0 & (is.na(appResults$prediction$comparisonCohortStartDate)),]
 
-    Noncases <- round(sum(predictionNonCases$value), 0)
-    cases <- round(sum(predictionCases$value), 0)
+    nonCases <- round(sum(predictionNonCases$value, na.rm = TRUE), 0)
+    cases <- round(sum(predictionCases$value, na.rm = TRUE), 0)
 
     # add other parameters to the input settings list
     appResults$PheValuator$inputSetting$phenotype <- phenotype
@@ -435,7 +511,7 @@
     appResults$PheValuator$inputSetting$minimumOffsetFromEnd <- minimumOffsetFromEnd
     appResults$PheValuator$inputSetting$excludeModelFromEvaluation <- as.character(excludeModelFromEvaluation)
 
-    appResults$PheValuator$diagnostics$Noncases <- Noncases
+    appResults$PheValuator$diagnostics$nonCases <- nonCases
     appResults$PheValuator$diagnostics$cases <- cases
 
     appResults$PheValuator$diagnostics$count30And70pct <- count30And70pct
@@ -451,7 +527,12 @@
 
     if(nrow(fullTestCases) > 0) { #no test cases found
       appResults$PheValuator$testCases <- as.data.frame(fullTestCases)
-      appResults$PheValuator$testCaseCovariates <- subjectCovariates[,c(11,12,9,6,7)]
+      if(ncol(subjectCovariates) == 12) { #rapid method
+        appResults$PheValuator$testCaseCovariates <- subjectCovariates[,c(11,12,9,6,7)]
+      } else { #predictPlp method
+        appResults$PheValuator$testCaseCovariates <- subjectCovariates[,c(8,9,6,7,4)]
+        names(appResults$PheValuator$testCaseCovariates)[4] <- "coeffValue"
+      }
     } else {
       appResults$PheValuator$testCases <- NULL
       appResults$PheValuator$testCaseCovariates <- NULL
@@ -475,6 +556,7 @@
 
     if(nrow(data.frame(appResults$PheValuator$testCases)) > 0) {
       df <- cbind(df, data.frame(appResults$PheValuator$testCases[,c(1,4,5,7:9,14:16)]))
+      names(df)[names(df)=="gender"] <- "genderSubject"
       colnames(df) <- SqlRender::camelCaseToSnakeCase(colnames(df))
       write.csv(df, file.path(exportFolder, "pv_test_subjects.csv"), row.names = FALSE)
 
@@ -499,13 +581,15 @@
 
 
     # remove temp cohort table
-    sql <- SqlRender::loadRenderTranslateSql(
-      sqlFilename = "DropTempTable.sql",
-      packageName = "PheValuator",
-      dbms = connectionDetails$dbms,
-      work_database_schema = workDatabaseSchema,
-      test_cohort = testCohort
-    )
-    DatabaseConnector::executeSql(connection = connection, sql = sql, progressBar = FALSE, reportOverallTime = FALSE)
+    if(tempTableCreated) {
+      sql <- SqlRender::loadRenderTranslateSql(
+        sqlFilename = "DropTempTable.sql",
+        packageName = "PheValuator",
+        dbms = connectionDetails$dbms,
+        work_database_schema = workDatabaseSchema,
+        test_cohort = testCohort
+      )
+      DatabaseConnector::executeSql(connection = connection, sql = sql, progressBar = FALSE, reportOverallTime = FALSE)
+    }
   }
 }